Use global constraints as much as possible, and clean up

+++ multidimensional_knapsack/src/check_solution.cpp
@@ -1,9 +1,9 @@
 #include "check_solution.hpp"
 
 bool check_solution( const std::vector<int>& solution,
-                     const std::vector<std::vector<int>>& coefficients,
+                     const std::vector<std::vector<double>>& coefficients,
                      const std::vector<int>& capacities,
-                     const std::vector<int>& values,
+                     const std::vector<double>& values,
                      int optimal )
 {
 	bool success = true;
